# api/openapi.yaml
openapi: 3.0.0

info:
  title: Telegram OIDC Provider API
  version: 0.1.0
  description: API for managing clients and users in the Telegram OIDC Provider

tags:
  - name: public
    description: Public API for browser-based requests (e.g. OIDC endpoints)
  - name: private
    description: Private API for internal use only

components:
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        message:
          type: string
          description: A brief error message
          example: Invalid request data
        details:
          description: Additional details about the error
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ObjectNotFoundDetails"
            - $ref: "#/components/schemas/ConflictDetails"
            - $ref: "#/components/schemas/ObjectInvalidDetails"
          discriminator:
            propertyName: type

    ObjectNotFoundDetails:
      type: object
      required: [type, object]
      properties:
        type:
          type: string
          enum: [object_not_found]
          description: Error detail type discriminator
        object:
          type: string
          description: The type of object that was not found
          example: client
        id:
          type: string
          description: The identifier of the object that was not found
          example: "123e456"
          nullable: true

    ConflictDetails:
      type: object
      required: [type, object]
      properties:
        type:
          type: string
          enum: [conflict]
          description: Error detail type discriminator
        object:
          type: string
          description: The type of object that was not found
          example: client
        feature:
          type: string
          description: The feature that caused the conflict
          example: redirect_uri

    ObjectInvalidDetails:
      type: object
      required: [type, object, field]
      properties:
        type:
          type: string
          enum: [object_invalid]
          description: Error detail type discriminator
        object:
          type: string
          description: The type of object that has invalid data
          example: bot
        field:
          type: string
          description: The field that contains invalid data
          example: token
        reason:
          type: string
          description: The reason why the field is invalid
          example: Invalid Telegram bot token format
          nullable: true

    BotResponse:
      type: object
      required: [id, name, username, url]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        client_id:
          type: string
          description: OIDC client ID associated with the bot
          example: "123e4567-e89b-12d3-a456-426614174000"
        photo_url:
          type: string
          format: url
          nullable: true
        url:
          type: string
          format: url

    BotBriefResponse:
      type: object
      required: [id, name, username]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string

paths:
  /bots:
    post:
      tags: [private]
      summary: Sync Telegram bot by token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 37
                  description: Telegram bot token
                  example: "123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11"

      responses:
        200:
          description: Bot synced successfully
          headers:
            Location:
              required: true
              description: Full URL of the Telegram bot resource.
              schema:
                type: string
                format: uri
                example: "https://example.com/bots/123"
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: integer
                    format: int64
                  updated:
                    type: boolean
                    description: Indicates whether the bot was updated (true) or not (false)
                  last_synced_at:
                    type: string
                    format: date-time
                    description: Timestamp of the last successful sync with Telegram

        201:
          description: Bot created successfully
          headers:
            Location:
              required: true
              description: Full URL of the Telegram bot resource.
              schema:
                type: string
                format: uri
                example: "https://example.com/bots/123"
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: integer
                    format: int64
        400:
          description: Invalid request data (e.g., invalid Telegram bot token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidBotTokenMalformed:
                  summary: Malformed Telegram bot token
                  value:
                    code: "object_invalid"
                    message: "bot has invalid field 'token': malformed"
                    details:
                      object: "bot"
                      field: "token"
                      reason: "malformed"
                invalidBotToken:
                  summary: Invalid Telegram bot token
                  value:
                    code: "object_invalid"
                    error: "bot has invalid field 'token'"
                    details:
                      object: "bot"
                      field: "token"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unexpected:
                  summary: Unexpected internal error
                  value:
                    code: "unexpected"
                    message: "unexpected error occurred"

  /widget/callback:
    get:
      tags: [public]
      summary: Login user by telegram widget auth data
      parameters:
        - in: query
          name: login_challenge
          required: true
          description: Unique login request identifier issued by ORY Hydra.
          schema:
            type: string
        - in: query
          name: telegram_widget_auth_data
          required: true
          description: >
            Collection of parameters returned by the Telegram Login Widget.
            Passed as form-style query parameters (explode=true), for example:
              - id
              - first_name
              - last_name
              - username
              - photo_url
              - auth_date
              - hash

            The server must:
              1. Verify the signature (hash) according to Telegram’s algorithm.
              2. Validate auth_date freshness.
              3. Accept or reject the login request via ORY Hydra.
          schema:
            type: object
            additionalProperties: true
          style: form
          explode: true
        - in: header
          name: User-Agent
          required: false
          description: User agent of the client
          schema:
            type: string
            example: "Mozilla/5.0 (Windows NT 10.0; Win64; x64)"
        - in: header
          name: Accept-Language
          required: false
          description: Optional language preference (e.g., en, ru, fr)
          schema:
            type: string
            example: "en-US,en;q=0.9"

      responses:
        203:
          description: Login flow processed. The client must follow the URL provided in the Location header.
          headers:
            Location:
              description: >
                Redirect URL generated by ORY Hydra.
                The client must redirect the user to this URL to continue
                the OIDC flow. It may contain either success parameters
                or error parameters.
              schema:
                type: string
                format: uri

  /miniapp/callback:
    get:
      tags: [public]
      summary: Login user by telegram mini app auth data
      parameters:
        - in: query
          name: login_challenge
          required: true
          description: Unique login request identifier issued by ORY Hydra.
          schema:
            type: string
        - in: query
          name: telegram_mini_app_auth_data
          required: true
          description: >
            Collection of parameters returned by the Telegram Mini App.
            Passed as form-style query parameters (explode=true), for example:
              - id
              - first_name
              - last_name
              - username
              - photo_url
              - auth_date
              - hash

            The server must:
              1. Verify the signature (hash) according to Telegram’s algorithm.
              2. Validate auth_date freshness.
              3. Accept or reject the login request via ORY Hydra.
          schema:
            type: object
            additionalProperties: true
          style: form
          explode: true
        - in: header
          name: User-Agent
          required: false
          description: User agent of the client
          schema:
            type: string
            example: "Mozilla/5.0 (Windows NT 10.0; Win64; x64)"
        - in: header
          name: Accept-Language
          required: false
          description: Optional language preference (e.g., en, ru, fr)
          schema:
            type: string
            example: "en-US,en;q=0.9"

      responses:
        203:
          description: Login flow processed. The client must follow the URL provided in the Location header.
          headers:
            Location:
              description: >
                Redirect URL generated by ORY Hydra.
                The client must redirect the user to this URL to continue
                the OIDC flow. It may contain either success parameters
                or error parameters.
              schema:
                type: string
                format: uri